feat: add plena wallet integration
