+++ .gitignore
@@ -36,3 +36,5 @@ yarn-error.log*
 
 # typechain
 # /typechain
+
+.env
