+++ package.json
@@ -13,6 +13,7 @@
         "got": "^12.1.0"
     },
     "dependencies": {
+        "plena-wallet-sdk": "^0.5.0",
         "@ethersproject/providers": "^5.6.8",
         "@reduxjs/toolkit": "^1.8.2",
         "@types/react-sparklines": "^1.7.2",
