+++ src/Components/WalletModals/config.ts
@@ -1,7 +1,8 @@
-import { walletStrings } from "../../wallet/constants";
+import { walletStrings } from "../../wallet/constant";
 import {
     BitKeep,
     Metamask,
+    Plena,
     SafePal,
     TrustWallet,
     WalletConnect,
@@ -12,6 +13,11 @@ export const walletConfig = [
         title: "Metamask",
         icon: Metamask,
         walletName: walletStrings.metamask,
+    },
+    {
+        title: "Plena",
+        icon: Plena,
+        walletName: walletStrings.plena,
     },
     {
         title: "Bitkeep",
