+++ src/Components/svgIcons/index.ts
@@ -1,4 +1,5 @@
 export * from "./Metamask";
+export * from "./Plena";
 export * from "./SafePal";
 export * from "./TrustWallet";
 export * from "./WalletConnect";
