+++ src/pages/_app.tsx
@@ -1,7 +1,14 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import AppProviders from "../AppProviders";
-import Layout from "../Components/Layout";
+import dynamic from "next/dynamic";
+
+const AppProviders = dynamic(() => import("../AppProviders/index"), {
+    ssr: false,
+});
+
+const Layout = dynamic(() => import("../Components/Layout/index"), {
+    ssr: false,
+});
 
 function MyApp({ Component, pageProps }: AppProps) {
     return (
