+++ src/pages/index.tsx
@@ -2,9 +2,7 @@ import type { NextPage } from "next";
 import Insurance from "./insurance";
 
 const Home: NextPage = () => {
-    return (
-        <Insurance />
-    );
+    return <Insurance />;
 };
 
 export default Home;
