+++ src/wallet/constant.ts
@@ -1,5 +1,6 @@
 export const walletStrings = {
     metamask: "metamask",
+    plena: "plena",
     bitkeep: "bitkeep",
     trustwallet: "trustwallet",
     safepal: "safepal",
