+++ src/wallet/hooks/useWallet.ts
@@ -1,20 +1,22 @@
 import { AbstractConnector } from "@web3-react/abstract-connector";
 import { bitKeep, injected, walletConnect } from "../connnectors";
-import { UnsupportedChainIdError, useWeb3React } from "@web3-react/core";
+import { useWeb3React } from "@web3-react/core";
 import {
     getSupportedChainsName,
     supportedChainIds,
 } from "../../constants/chainIds";
-import { setupNetwork, getConnectionError } from "../utils";
+import { setupNetwork } from "../utils";
 import { useCallback, useEffect } from "react";
 import CustomToast, { STATUS, TYPE } from "../../Components/CustomToast";
 import { toast } from "react-toastify";
 import { getChainId } from "../../utils/helpers";
-import { walletLocalStorageKey, walletStrings } from "../constants";
+import { walletLocalStorageKey, walletStrings } from "../constant";
 import { isMobile, isDesktop } from "react-device-detect";
+import { usePlenaWallet } from "plena-wallet-sdk";
 
 const useWallet = () => {
     const { activate, deactivate } = useWeb3React();
+    const { openModal } = usePlenaWallet();
 
     const _ethereumListener = async () =>
         new Promise<void>((resolve) =>
@@ -97,7 +99,11 @@ const useWallet = () => {
     }, []);
 
     const connectWallet = useCallback(
-        async (name: string) => {
+        async (name: string, isPlena: boolean = false) => {
+            if (isPlena) {
+                openModal();
+                return window.localStorage.setItem(walletLocalStorageKey, name);
+            }
             let connector: AbstractConnector;
             let walletName = name;
             const injectedWallets = [
@@ -191,7 +197,7 @@ const useWallet = () => {
                 toast(body);
             }
         },
-        [activate]
+        [activate, openModal]
     );
 
     const disconnectWallet = () => {
