+++ types.d.ts
@@ -1,0 +1,2 @@
+declare module "plena-wallet-sdk";
+declare module "swr/immutable";
